REDSHOP-4606 Entity optimize
